# src/backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Фильм на вечер API
  description: API для приложения управления коллекциями фильмов
  version: 1.0.0
servers:
  - url: /api
    description: Основной API сервер
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Побег из Шоушенка"
        description:
          type: string
          example: "Двое заключенных находят утешение и искупление через акты порядочности..."
        image_url:
          type: string
          example: "https://example.com/movie.jpg"
        image_base64:
          type: string
          nullable: true
        rating:
          type: number
          format: float
          example: 9.3
        genre:
          type: string
          example: "Драма"
        duration:
          type: integer
          example: 142
        year:
          type: integer
          example: 1994
        loaded_from_tmdb:
          type: boolean
          example: true
      required:
        - id
        - title
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Неверные данные запроса"
        details:
          type: object
          properties:
            errors:
              type: string
              example: "Отсутствует обязательное поле: name"
      required:
        - code
        - message
...
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          example: "Шоушенк"
        page:
          type: integer
          example: 1
          default: 1
      required:
        - query
    AISearchRequest:
      type: object
      properties:
        query:
          type: string
          example: "Фильм о побеге из тюрьмы"
      required:
        - query
    AISearchResponse:
      type: object
      properties:
        film_name:
          type: string
          example: "Побег из Шоушенка"
      required:
        - film_name
security:
  - BearerAuth: []
...
paths:
...
  /movie/search:
    post:
      tags:
        - Поиск
      summary: Поиск фильмов
      description: Поиск фильмов по названию или другим критериям
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movie/search/ai:
    post:
      tags:
        - Поиск
      summary: Поиск фильмов с помощью ИИ
      description: Поиск фильмов с использованием обработки естественного языка
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AISearchRequest'
      responses:
        '200':
          description: Результат поиска с помощью ИИ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISearchResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'